% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_Individual_Analysis.R
\name{AI_Individual_Analysis}
\alias{AI_Individual_Analysis}
\title{Ancestry-informed individual-variant analysis using score test}
\usage{
AI_Individual_Analysis(
  chr,
  individual_results,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  geno_missing_imputation = c("mean", "minor"),
  tol = .Machine$double.eps^0.25,
  max_iter = 1000,
  SPA_p_filter = TRUE,
  p_filter_cutoff = 0.05,
  find_weight = TRUE
)
}
\arguments{
\item{chr}{chromosome.}

\item{individual_results}{the data frame of (significant) individual variants of interest for ancestry-informed analysis.
The first 4 columns should correspond to chromosome (CHR), position (POS), reference allele (REF), and alternative allele (ALT).}

\item{start_loc}{starting location (position) of the genetic region for each individual variant to be analyzed using score test.}

\item{end_loc}{ending location (position) of the genetic region for each individual variant to be analyzed using score test.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function with two or more specified ancestries in \code{pop.groups},
or the output from \code{\link{fit_nullmodel}} function transformed using the \code{\link{staar2aistaar_nullmodel}} function.}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "variant", "SNV", or "Indel" (default = "variant").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{tol}{a positive number specifying tolerance, the difference threshold for parameter
estimates in saddlepoint approximation algorithm below which iterations should be stopped (default = ".Machine$double.eps^0.25").}

\item{max_iter}{a positive integer specifying the maximum number of iterations for applying the saddlepoint approximation algorithm (default = "1000").}

\item{SPA_p_filter}{logical: are only the variants with a score-test-based p-value smaller than a pre-specified threshold use the SPA method to recalculate the p-value, only used for imbalanced case-control setting (default = TRUE).}

\item{p_filter_cutoff}{threshold for the p-value recalculation using the SPA method, only used for imbalanced case-control setting (default = 0.05)}

\item{find_weight}{logical: should the ancestry group-specific weights and weighting scenario-specific p-values for each base test be saved as output (default = FALSE).}
}
\value{
A data frame containing the score test p-value and the estimated effect size of the minor allele for each individual variant in the given genetic region.
The first 4 columns correspond to chromosome (CHR), position (POS), reference allele (REF), and alternative allele (ALT).
If \code{find_weight} is TRUE, returns a list containing the ancestry-informed score test p-values, estimated effect sizes with corresponding variant characteristics,
as well as the ensemble weights under two sampling scenarios and p-values under scenarios 1, 2, and combined for each base test.
}
\description{
The \code{AI_Individual_Analysis} function takes in chromosome, an user-defined variant list,
the object of opened annotated GDS file, and the object from fitting the null model to analyze the association between a
quantitative/dichotomous phenotype and each individual variant by using score test.
The results of the ancestry-informed analysis correspond to ensemble p-values across base tests,
with the option to return a list of base weights and p-values for each base test.
}
\references{
Chen, H., et al. (2016). Control for population structure and relatedness for binary traits
in genetic association studies via logistic mixed models. \emph{The American Journal of Human Genetics}, \emph{98}(4), 653-666.
(\href{https://doi.org/10.1016/j.ajhg.2016.02.012}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting
noncoding rare-variant associations of large-scale whole-genome sequencing
studies. \emph{Nature Methods}, \emph{19}(12), 1599-1611.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})
}
