% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_Window_SCANG.R
\name{Dynamic_Window_SCANG}
\alias{Dynamic_Window_SCANG}
\title{Genetic region analysis of dynamic windows using SCANG-STAAR procedure}
\usage{
Dynamic_Window_SCANG(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  Lmin = 40,
  Lmax = 300,
  steplength = 10,
  rare_maf_cutoff = 0.01,
  p_filter = 1e-08,
  f = 0,
  alpha = 0.1,
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(TRUE, FALSE),
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{chromosome.}

\item{start_loc}{starting location (position) of the genetic region to be analyzed using SCANG-STAAR procedure.}

\item{end_loc}{ending location (position) of the genetic region to be analyzed using SCANG-STAAR procedure.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is the output from \code{\link{fit_nullmodel}} function
and transformed using the \code{\link{staar2scang_nullmodel}} function.}

\item{Lmin}{minimum number of variants in searching windows (default = 40).}

\item{Lmax}{maximum number of variants in searching windows (default = 300).}

\item{steplength}{difference of number of variants in searching windows, that is, the number of variants in
searching windows are Lmin, Lmin+steplength, Lmin+steplength,..., Lmax (default = 10).}

\item{rare_maf_cutoff}{a cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{p_filter}{a filtering threshold of screening method for SKAT in SCANG-STAAR. SKAT p-values are calculated for regions whose p-value
is possibly smaller than the filtering threshold (default = 1e-8).}

\item{f}{an overlap fraction, which controls for the overlapping proportion of of detected regions. For example,
when f=0, the detected regions are non-overlapped with each other,
and when f=1, we keep every susceptive region as detected regions (default = 0).}

\item{alpha}{family-wise/genome-wide significance level (default = 0.1).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "SNV", "Indel", or "variant" (default = "SNV").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file \cr (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{Use_annotation_weights}{use annotations as weights or not (default = TRUE).}

\item{Annotation_name}{a vector of annotation names used in SCANG-STAAR (default = NULL).}

\item{silent}{logical: should the report of error messages be suppressed (default = FALSE).}
}
\value{
The function returns a list with the following members:

\code{SCANG_O_res}: A matrix that summarizes the significant region detected by SCANG-STAAR-O,
including the negative log transformation of SCANG-STAAR-O p-value ("-logp"), chromosome ("chr"), start position ("start_pos"), end position ("end_pos"),
family-wise/genome-wide error rate (GWER) and the number of variants  ("SNV_num").

\code{SCANG_O_top1}: A vector of length 4 which summarizes the top 1 region detected by SCANG-STAAR-O.
including the negative log transformation of SCANG-STAAR-O p-value ("-logp"), chromosome ("chr"), start position ("start_pos"), end position ("end_pos"),
family-wise/genome-wide error rate (GWER) and the number of variants  ("SNV_num").

\code{SCANG_O_emthr}: A vector of Monte Carlo simulation sample for generating the empirical threshold. The 1-alpha quantile of this vector is
the empirical threshold.

\code{SCANG_S_res, SCANG_S_top1, SCANG_S_emthr}: Analysis
results using SCANG-STAAR-S. Details see SCANG-STAAR-O.

\code{SCANG_B_res, SCANG_B_top1, SCANG_B_emthr}: Analysis
results using SCANG-STAAR-B. Details see SCANG-STAAR-O.
}
\description{
The \code{Dynamic_Window_SCANG} function takes in chromosome, starting location, ending location,
the object of opened annotated GDS file, and the object from fitting the null model to analyze the association between a
quantitative/dichotomous phenotype and variants in a genetic region by using SCANG-STAAR procedure.
For each dynamic window, the scan statistic of SCANG-STAAR-O is the set-based p-value of an omnibus test that aggregated p-values
across different types of multiple annotation-weighted variant-set tests SKAT(1,1), SKAT(1,25), Burden(1,1) and Burden(1,25) using ACAT method;
the scan statistic of SCANG-STAAR-S is the set-based p-value of STAAR-S, which is an omnibus test that aggregated p-values
across multiple annotation-weighted variant-set tests SKAT(1,1) and SKAT(1,25) using ACAT method;
the scan statistic of SCANG-STAAR-B is the set-based p-value of STAAR-B, which is an omnibus test that aggregated p-values
across multiple annotation-weighted variant-set tests Burden(1,1) and Burden(1,25) using ACAT method.
}
\references{
Li, Z., Li, X., et al. (2019). Dynamic scan procedure for
detecting rare-variant association regions in whole-genome sequencing studies.
\emph{The American Journal of Human Genetics}, \emph{104}(5), 802-814.
(\href{https://doi.org/10.1016/j.ajhg.2019.03.002}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Human Genetics}, \emph{104}(3), 410-421.
(\href{https://doi.org/10.1016/j.ajhg.2019.01.002}{pub})
}
