% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis2staar_nullmodel.R
\name{genesis2staar_nullmodel}
\alias{genesis2staar_nullmodel}
\title{Transforming the null model object fitted using GENESIS to the null model object to be used for STAAR}
\usage{
genesis2staar_nullmodel(obj_nullmodel_genesis)
}
\arguments{
\item{obj_nullmodel_genesis}{an object from fitting the null model, which is the
output from \code{fitNullModel} function in the \code{GENESIS} package.}
}
\value{
an object from fitting the null model for related samples to be used for STAAR procedure,
which is the output from \code{\link{fit_nullmodel}} function.
}
\description{
The \code{genesis2staar_nullmodel} function takes in the object from fitting the null model using the \code{GENESIS}
package and transforms it to the object from fitting the null model to be used for STAAR procedure.
}
\references{
Gogarten, S.M., Sofer, T., Chen, H., et al. (2019). Genetic association testing using the
GENESIS R/Bioconductor package. \emph{Bioinformatics}, \emph{35}(24), 5346-5348.
(\href{https://doi.org/10.1093/bioinformatics/btz567}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting
noncoding rare-variant associations of large-scale whole-genome sequencing
studies. \emph{Nature Methods}, \emph{19}(12), 1599-1611.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})
}
