% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Analysis.R
\name{Individual_Analysis}
\alias{Individual_Analysis}
\title{Individual-variant analysis using score test}
\usage{
Individual_Analysis(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  mac_cutoff = 20,
  subset_variants_num = 5000,
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  geno_missing_imputation = c("mean", "minor")
)
}
\arguments{
\item{chr}{chromosome.}

\item{start_loc}{starting location (position) of the genetic region for each individual variant to be analyzed using score test.}

\item{end_loc}{ending location (position) of the genetic region for each individual variant to be analyzed using score test.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{mac_cutoff}{the cutoff of minimum minor allele count in
defining individual variants (default = 20).}

\item{subset_variants_num}{the number of variants to run per subset for each time (default = 5e3).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "variant", "SNV", or "Indel" (default = "variant").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}
}
\value{
a data frame containing the score test p-value and the estimated effect size for each individual variant in the given genetic region.
The first 4 columns correspond to chromosome (CHR), position (POS), reference allele (REF), and alternative allele (ALT).
}
\description{
The \code{Individual_Analysis} function takes in chromosome, starting location, ending location,
the object of opened annotated GDS file, and the object from fitting the null model to analyze the association between a
quantitative/dichotomous phenotype and each individual variant in a genetic region by using score test.
}
\references{
Chen, H., et al. (2016). Control for population structure and relatedness for binary traits
in genetic association studies via logistic mixed models. \emph{The American Journal of Human Genetics}, \emph{98}(4), 653-666.
(\href{https://doi.org/10.1016/j.ajhg.2016.02.012}{pub})
}
