% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Analysis_cond.R
\name{Individual_Analysis_cond}
\alias{Individual_Analysis_cond}
\title{Individual-variant conditional analysis using score test for whole-genome sequencing data}
\usage{
Individual_Analysis_cond(
  chr,
  individual_results,
  genofile,
  obj_nullmodel,
  known_loci = NULL,
  method_cond = c("optimal", "naive"),
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  geno_missing_imputation = c("mean", "minor")
)
}
\arguments{
\item{chr}{chromosome.}

\item{individual_results}{the data frame of the significant individual variants for conditional analysis using score test.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT) (default = NULL).}

\item{method_cond}{a character value indicating the method for conditional analysis.
\code{optimal} refers to regressing residuals from the null model on \code{known_loci}
as well as all covariates used in fitting the null model (fully adjusted) and taking the residuals;
\code{naive} refers to regressing residuals from the null model on \code{known_loci}
and taking the residuals (default = \code{optimal}).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{variants include in the analysis. Choices include "variant", "SNV", or "Indel" (default = "variant").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}
}
\value{
a data frame containing the conditional score test p-value and effect size for each significant individual variant in the given set.
}
\description{
The \code{Individual_Analysis_cond} function takes in chromosome, starting location, ending location,
the object of opened annotated GDS file, the object from fitting the null model,
and the set of known variants to be adjusted for in conditional analysis to analyze the conditional association between a
quantitative/dichotomous phenotype and each significant individual variant by using score test.
}
\references{
Chen, H., et al. (2016). Control for population structure and relatedness for binary traits
in genetic association studies via logistic mixed models. \emph{The American Journal of Human Genetics}, \emph{98}(4), 653-666.
(\href{https://doi.org/10.1016/j.ajhg.2016.02.012}{pub})

Sofer, T., et al. (2019). A fully adjusted two-stage procedure for rank-normalization
in genetic association studies. \emph{Genetic Epidemiology}, \emph{43}(3), 263-275.
(\href{https://doi.org/10.1002/gepi.22188}{pub})
}
