% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_pruning.R
\name{LD_pruning}
\alias{LD_pruning}
\title{Linkage disequilibrium (LD) pruning procedure}
\usage{
LD_pruning(
  chr,
  genofile,
  obj_nullmodel,
  variants_list,
  maf_cutoff = 0.01,
  cond_p_thresh = 1e-04,
  method_cond = c("optimal", "naive"),
  QC_label = "annotation/filter",
  variant_type = c("variant", "SNV", "Indel"),
  geno_missing_imputation = c("mean", "minor"),
  geno_position_ascending = TRUE
)
}
\arguments{
\item{chr}{chromosome.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{variants_list}{the data frame of variants to be LD-pruned in sequential conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT).}

\item{maf_cutoff}{the cutoff of minimum minor allele frequency in
defining individual variants to be LD-pruned (default = 0.01).}

\item{cond_p_thresh}{the cutoff of maximum conditional p-value allowed for variants to be kept in the LD-pruned
list of variants (default = 1e-04).}

\item{method_cond}{a character value indicating the method for conditional analysis.
\code{optimal} refers to regressing residuals from the null model on \code{known_loci}
as well as all covariates used in fitting the null model (fully adjusted) and taking the residuals;
\code{naive} refers to regressing residuals from the null model on \code{known_loci}
and taking the residuals (default = \code{optimal}).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "variant", "SNV", or "Indel" (default = "variant").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{geno_position_ascending}{logical: are the variant positions in ascending order in the GDS/aGDS file (default = TRUE).}
}
\value{
a data frame containing the list of LD-pruned variants in the given chromosome.
}
\description{
The \code{LD_pruning} function takes in chromosome, the object of opened annotated GDS file,
the object from fitting the null model, and a given list of variants to perform LD pruning
among these variants in sequential conditional analysis by using score test.
}
