% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncRNA.R
\name{ncRNA}
\alias{ncRNA}
\title{Gene-centric analysis of noncoding RNA category using STAAR procedure}
\usage{
ncRNA(
  chr,
  gene_name,
  genofile,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(TRUE, FALSE),
  Annotation_name = NULL,
  silent = FALSE
)
}
\arguments{
\item{chr}{chromosome.}

\item{gene_name}{name of the ncRNA gene to be analyzed using STAAR procedure.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{type of variant included in the analysis. Choices include "SNV", "Indel", or "variant" (default = "SNV").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{Use_annotation_weights}{use annotations as weights or not (default = TRUE).}

\item{Annotation_name}{annotations used in STAAR (default = NULL).}

\item{silent}{logical: should the report of error messages be suppressed (default = FALSE).}
}
\value{
a data frame of STAAR p-values (including STAAR-O) corresponding to the exonic and splicing category of the given ncRNA gene.
}
\description{
The \code{ncRNA} function takes in chromosome, gene name,
the object of opened annotated GDS file, and the object from fitting the null model to analyze the association between a
quantitative/dichotomous phenotype and the exonic and splicing category of an ncRNA gene by using STAAR procedure.
For each ncRNA category, the STAAR-O p-value is a p-value from an omnibus test
that aggregated SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25),
and ACAT-V(1,1) together with p-values of each test weighted by each annotation
using Cauchy method.
}
\references{
Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})
}
