% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sliding_Window_cond.R
\name{Sliding_Window_cond}
\alias{Sliding_Window_cond}
\title{Genetic region conditional analysis of sliding windows using STAAR procedure}
\usage{
Sliding_Window_cond(
  chr,
  start_loc,
  end_loc,
  genofile,
  obj_nullmodel,
  known_loci,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  method_cond = c("optimal", "naive"),
  QC_label = "annotation/filter",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(TRUE, FALSE),
  Annotation_name = NULL
)
}
\arguments{
\item{chr}{chromosome.}

\item{start_loc}{starting location (position) of the sliding window to be analyzed using STAAR procedure.}

\item{end_loc}{ending location (position) of the sliding window to be analyzed using STAAR procedure.}

\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{\link{genesis2staar_nullmodel}} function.}

\item{known_loci}{the data frame of variants to be adjusted for in conditional analysis and should
contain 4 columns in the following order: chromosome (CHR), position (POS), reference allele (REF),
and alternative allele (ALT).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{method_cond}{a character value indicating the method for conditional analysis.
\code{optimal} refers to regressing residuals from the null model on \code{known_loci}
as well as all covariates used in fitting the null model (fully adjusted) and taking the residuals;
\code{naive} refers to regressing residuals from the null model on \code{known_loci}
and taking the residuals (default = \code{optimal}).}

\item{QC_label}{channel name of the QC label in the GDS/aGDS file (default = "annotation/filter").}

\item{variant_type}{variants include in the analysis. Choices include "variant", "SNV", or "Indel" (default = "SNV").}

\item{geno_missing_imputation}{method of handling missing genotypes. Either "mean" or "minor" (default = "mean").}

\item{Annotation_dir}{channel name of the annotations in the aGDS file  (default = "annotation/info/FunctionalAnnotation").}

\item{Annotation_name_catalog}{a data frame containing the name and the corresponding channel name in the aGDS file.}

\item{Use_annotation_weights}{use annotations as weights or not (default = TRUE).}

\item{Annotation_name}{annotations used in STAAR.}
}
\value{
a data frame containing the conditional STAAR p-values (including STAAR-O) corresponding to the sliding window in the given genetic region.
}
\description{
The \code{Sliding_Window_cond} function takes in chromosome, starting location, ending location,
the object of opened annotated GDS file, the object from fitting the null model,
and the set of known variants to be adjusted for in conditional analysis to analyze the conditional association between a
quantitative/dichotomous phenotype and variants in a genetic region by using STAAR procedure.
For each sliding window, the conditional STAAR-O p-value is a p-value from an omnibus test
that aggregated conditional SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25),
and ACAT-V(1,1) together with conditional p-values of each test weighted by each annotation
using Cauchy method.
}
\references{
Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})

Sofer, T., et al. (2019). A fully adjusted two-stage procedure for rank-normalization
in genetic association studies. \emph{Genetic Epidemiology}, \emph{43}(3), 263-275.
(\href{https://doi.org/10.1002/gepi.22188}{pub})
}
