% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staar2scang_nullmodel.R
\name{staar2scang_nullmodel}
\alias{staar2scang_nullmodel}
\title{Transforming the null model object fitted using STAAR to the null model object to be used for SCANG-STAAR}
\usage{
staar2scang_nullmodel(obj_nullmodel)
}
\arguments{
\item{obj_nullmodel}{an object from fitting the null model, which is either the output from \code{\link{fit_nullmodel}} function,
or the output from \code{fitNullModel} function in the \code{GENESIS} package and transformed using the \code{genesis2staar_nullmodel} function.}
}
\value{
An object from fitting the null model for related samples to be used for SCANG-STAAR procedure,
which is the output from \code{fit_null_glmmkin_SCANG} function for related samples in the \code{SCANG} package.
}
\description{
The \code{staar2scang_nullmodel} function takes in the object from fitting the null model
and transforms it to the object from fitting the null model to be used for SCANG-STAAR procedure.
}
\references{
Li, Z., Li, X., et al. (2022). A framework for detecting
noncoding rare-variant associations of large-scale whole-genome sequencing
studies. \emph{Nature Methods}, \emph{19}(12), 1599-1611.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})

Li, Z., Li, X., et al. (2019). Dynamic scan procedure for
detecting rare-variant association regions in whole-genome sequencing studies.
\emph{The American Journal of Human Genetics}, \emph{104}(5), 802-814.
(\href{https://doi.org/10.1016/j.ajhg.2019.03.002}{pub})
}
